+++ .eslintrc.cjs
@@ -3,15 +3,16 @@ require("@rushstack/eslint-patch/modern-module-resolution")
 
 module.exports = {
   root: true,
-  "extends": [
+  extends: [
     "plugin:vue/vue3-essential", 
     "eslint:recommended",
     "@vue/eslint-config-typescript",
     "@vue/eslint-config-prettier/skip-formatting",
     "plugin:storybook/recommended"
   ],
+  ignorePatterns: ["*.config.ts", "*.spec.ts"],
   parserOptions: {
-    project: "./tsconfig.app.json",
+    project: ["./tsconfig.app.json"],
     ecmaVersion: "latest"
   },
   plugins: ["import", "import-access"],
@@ -41,14 +42,9 @@ module.exports = {
       {
         zones: [
           {
-            target: "./src/components",
-            from: "./src/lib/infra",
-            message: "NG!!!",
-          },
-          {
-            target: "./src/views",
-            from: "./src/views/!(_shared_|components)/**/*",
-            message: "not shared",
+            target: "./src/components/parts",
+            from: "./src/components/!(parts)",
+            message: "import禁止",
           },
         ],
       },
