+++ src/components/parts/box/index.ts
@@ -1,4 +1,5 @@
 import FlexBox from "./FlexBox.vue"
 import SpaceBox from "./SpaceBox.vue"
 
+/** @public */
 export { FlexBox, SpaceBox }
