+++ src/components/parts/button/index.ts
@@ -1,3 +1,4 @@
 import TheButton from "./TheButton.vue"
 
+/** @public */
 export { TheButton }
