+++ src/components/parts/form/index.ts
@@ -1,4 +1,5 @@
 import SelectBox from "./SelectBox.vue"
 import TextInput from "./TextInput.vue"
 
+/** @public */
 export { SelectBox, TextInput }
