+++ src/components/parts/icons/index.ts
@@ -1,3 +1,3 @@
 import GoogleFontIcon from "./GoogleFontIcon.vue"
-
+/** @public */
 export { GoogleFontIcon }
