+++ src/components/parts/table/index.ts
@@ -2,4 +2,5 @@ import TableContainer from "./TableContainer.vue"
 import TableHeader from "./TableHeader.vue"
 import TableRecord from "./TableRecord.vue"
 
+/** @public */
 export { TableContainer, TableHeader, TableRecord }
