+++ src/lib/model/auth/repository.ts
@@ -3,6 +3,8 @@ type LoginParams = {
   password: string
 }
 
+// TODO: visiblity検討
+/** @public */
 export interface AuthRepository {
   login(params: LoginParams): Promise<void>
 }
