+++ src/provider/app-provider/index.ts
@@ -1,3 +1,3 @@
 import AppProvider from "./AppProvider.vue"
-
+/** @public */
 export { AppProvider }
