+++ src/provider/auth-provider/index.ts
@@ -1,3 +1,3 @@
-import AuthProvider from "./AuthProvider.vue";
-
-export { AuthProvider };
+import AuthProvider from "./AuthProvider.vue"
+/** @public */
+export { AuthProvider }
