+++ src/provider/auth-provider/use-snackbar.ts
@@ -7,9 +7,9 @@ type Context = {
 }
 
 const key = Symbol() as InjectionKey<Context>
-
+/** @public */
 export const createContext = (context: Context) => {
   provide(key, context)
 }
-
+/** @public */
 export const useErrorHandler = () => inject(key)
