+++ src/views/_shared_/components/index.ts
@@ -1,3 +1,4 @@
 import ModalContainer from "./ModalContainer.vue"
 
+/** @public */
 export { ModalContainer }
