+++ src/views/_shared_/custom-validate/interface.ts
@@ -1,0 +1,5 @@
+import type { ZodType } from "zod"
+
+export interface CustomValidate {
+  get value(): ZodType
+}
