--- src/views/_shared_/interface.ts
@@ -1,5 +1,0 @@
-import type { ZodType } from "zod";
-
-export interface CustomValidate {
-  get value(): ZodType;
-}
