--- src/views/home/components/add-modal/index.ts
@@ -1,3 +1,0 @@
-import AddModal from "./AddModal.vue"
-
-export { AddModal }
