--- src/views/home/components/edit-modal/index.ts
@@ -1,2 +1,0 @@
-import EditModal from "./EditModal.vue"
-export { EditModal }
